Update hardhat-play submodule
This commit updates the hardhat-play submodule to the latest commit.

The hardhat-play submodule is used to run the Hardhat Playground.

The Hardhat Playground is a web-based IDE that allows you to write, compile,
deploy and interact with Solidity smart contracts.

The Hardhat Playground is available at https://play.hardhat.org.