Add a new NFT contract
This commit adds a new NFT contract to the project.

The new contract is called DynamicSvgNft and it is a simple NFT contract
that allows the owner to change the tokenURI of the NFT.

This is a very simple contract that is meant to be used as an example
of how to create a new NFT contract.